feat: use mmser datastructures for mmapability

+++ src/fmindex-collection/bitvector/PairedBitvector2L.h
@@ -60,7 +60,7 @@ struct PairedBitvector2L {
     template <std::ranges::sized_range range_t>
         requires std::same_as<std::ranges::range_value_t<range_t>, std::bitset<l1_bits_ct>>
     PairedBitvector2L(range_t&& _range) {
-        auto _length = _range.size()*l1_bits_ct;
+        auto _length = static_cast<size_t>(_range.size()*l1_bits_ct);
 
         l0.resize((_length+l0_bits_ct)/(l0_bits_ct*2) + 1);
         l1.resize((_length+l1_bits_ct)/(l1_bits_ct*2) + 1);
